Create tests for QueryWebSocketClient
